+++ src/main/java/com/models/entity/dao/Test.java
@@ -16,28 +16,27 @@ public class Test extends baseEntity implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @GenericGenerator(strategy = "identity", name = "id")
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "user_id", nullable = false)
-    private Integer userId;
+    private Long userId;
     @Column(name = "user_name")
     private String userName;
     @Column(name = "email")
     private String email;
 
-
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -56,7 +55,5 @@ public class Test extends baseEntity implements Serializable {
     public void setEmail(String email) {
         this.email = email;
     }
-
-
 }
